# swagger.yaml
openapi: 3.0.3
info:
    title: Coffee Supply Management System API
    version: "1.0.0"
    contact:
        name: Team 67 - PZSP2
        email: coffee-system@pw.edu.pl
...
paths:
...
    /orders:
        get:
            tags: [Orders]
            summary: Order history
            operationId: listOrders
            security:
                - BearerAuth: []
            parameters:
                - name: status
                  in: query
                  schema:
                      type: string
                      enum: [confirmed, pending, cancelled]
            responses:
                "200":
                    description: List of orders
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/OrderRecord"

        post:
            tags: [Orders]
            summary: Confirm order schedule
            operationId: confirmOrders
            security:
                - BearerAuth: []
            parameters:
                - name: result_id
                  in: query
                  required: true
                  schema:
                      type: string
                  description: Optimization result ID to confirm
            responses:
                "201":
                    description: Order confirmed
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/OrderRecord"
    /orders/{order_id}:
        get:
            tags: [Orders]
            summary: Get order details
            operationId: getOrder
            security:
                - BearerAuth: []
            parameters:
                - name: order_id
                  in: path
                  required: true
                  description: Order UUID
                  schema:
                      type: string
                      format: uuid
            responses:
                "200":
                    description: Order details
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/OrderRecord"
                "401":
                    $ref: "#/components/responses/Unauthorized"
                "404":
                    $ref: "#/components/responses/NotFound"
    /orders/{order_id}/status:
        patch:
            tags: [Orders]
            summary: Update order status
            operationId: updateOrderStatus
            security:
                - BearerAuth: []
            parameters:
                - name: order_id
                  in: path
                  required: true
                  description: Order UUID
                  schema:
                      type: string
                      format: uuid
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [status]
                            properties:
                                status:
                                    type: string
                                    enum: [confirmed, pending, cancelled]
                        example:
                            status: cancelled
            responses:
                "200":
                    description: Order status updated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/OrderRecord"
                "401":
                    $ref: "#/components/responses/Unauthorized"
                "404":
                    $ref: "#/components/responses/NotFound"
...
components:
    securitySchemes:
        BearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: Coordinator session token (from `/auth/login`)
        ApiKeyAuth:
            type: apiKey
            in: header
            name: X-Api-Key
            description: Distributor API key (format `cof_...`)
...
    responses:
        Unauthorized:
            description: Missing or invalid token/API key
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
                    example:
                        detail: "Authentication required"
        NotFound:
            description: Resource not found
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
                    example:
                        detail: "Not found"
    schemas:
        ErrorResponse:
            type: object
            properties:
                detail:
                    type: string
...
        OrderItem:
            type: object
            properties:
                distributor_id:
                    type: string
                    format: uuid
                building_id:
                    type: string
                    format: uuid
                day:
                    type: integer
                    minimum: 1
                threshold_level:
                    type: integer
                    minimum: 0
                    description: "0 = no tier, >= 1 = discount tier number"
                quantity_kg:
                    type: number
                    format: float
                    minimum: 0
...
        OrderRecord:
            type: object
            properties:
                id:
                    type: string
                result_id:
                    type: string
                scenario_id:
                    type: string
                orders:
                    type: array
                    items:
                        $ref: "#/components/schemas/OrderItem"
                total_cost_pln:
                    type: number
                confirmed_by:
                    type: string
                    format: UUID
                status:
                    type: string
                    enum: [confirmed, pending, cancelled]
                created_at:
                    type: string
                    format: date-time
                updated_at:
                    type: string
                    format: date-time
...